move and renamed dynamic files
