+++ database/editData/editRecipe.js
@@ -1,4 +1,4 @@
-import { client } from "../../client";
+import { client } from "../client";
 
 export default async function editRecipe(key, recipeId, recipeValue){
     const db = client.db('devdb')
